text2sql/requirements.txt
mysql-connector-python 
transformers 
torch